+++ tests/integration/opus.ts
@@ -1,8 +1,7 @@
+import * as assert from 'assert';
 import Call from '../../lib/twilio/call';
 import Device from '../../lib/twilio/device';
 import { generateAccessToken } from '../lib/token';
-import * as assert from 'assert';
-import { EventEmitter } from 'events';
 
 // (rrowland) The TwiML expected by these tests can be found in the README.md
 
@@ -28,15 +27,10 @@ describe('Opus', function() {
     device1 = new Device(token1, options);
     device2 = new Device(token2, options);
 
-    const deviceReadyPromise = Promise.all([
-      expectEvent(Device.EventName.Registered, device1),
-      expectEvent(Device.EventName.Registered, device2),
+    return Promise.all([
+      device1.register(),
+      device2.register(),
     ]);
-
-    device1.register();
-    device2.register();
-
-    return deviceReadyPromise;
   });
 
   after(() => {
@@ -52,27 +46,25 @@ describe('Opus', function() {
   });
 
   describe('device 1 calls device 2', () => {
-    before(done => {
-      device2.once(Device.EventName.Incoming, () => done());
-      (device1['connect'] as any)({
+    let call1: Call;
+    let call2: Call;
+
+    before(() => new Promise(async resolve => {
+      device2.once(Device.EventName.Incoming, (call: Call) => {
+        call2 = call;
+        resolve();
+      });
+      call1 = await (device1['connect'] as any)({
         params: { To: identity2, Custom1: 'foo + bar', Custom2: undefined, Custom3: '我不吃蛋' }
       });
-    });
+    }));
 
     describe('and device 2 accepts', () => {
-      let call1: Call;
-      let call2: Call;
 
       beforeEach(() => {
-        const conn1: Call | undefined | null = device1.activeCall || device1.calls[0];
-        const conn2: Call | undefined | null = device2.activeCall || device2.calls[0];
-
-        if (!conn1 || !conn2) {
+        if (!call1 || !call2) {
           throw new Error(`Calls weren't both open at beforeEach`);
         }
-
-        call1 = conn1;
-        call2 = conn2;
       });
 
       it('should connect the call', (done) => {
@@ -148,9 +140,3 @@ describe('Opus', function() {
     });
   });
 });
-
-function expectEvent(eventName: string, emitter: EventEmitter) {
-  return new Promise(resolve => {
-    emitter.once(eventName, () => resolve());
-  });
-}